@@ -32,19 +32,19 @@
   },
   "homepage": "https://github.com/purpl3mint/Ninja-Boilerplate#readme",
   "devDependencies": {
-    "@babel/core": "^7.14.8",
-    "@babel/preset-env": "^7.14.8",
-    "browser-sync": "^2.27.5",
-    "chrome-launcher": "^0.14.0",
+    "@babel/core": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
+    "browser-sync": "^2.27.10",
+    "chrome-launcher": "^0.15.1",
     "critical": "^4.0.1",
-    "del": "^6.0.0",
+    "del": "^7.0.0",
     "editorconfig-cli": "^0.3.0",
-    "eslint": "^7.32.0",
-    "eslint-config-htmlacademy": "^4.0.2",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint": "^8.20.0",
+    "eslint-config-htmlacademy": "^6.0.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-standard": "^5.0.0",
     "gulp": "^4.0.2",
     "gulp-autoprefixer": "^8.0.0",
@@ -53,32 +53,34 @@
     "gulp-connect": "^5.7.0",
     "gulp-eslint": "^6.0.0",
     "gulp-html-bem-validator": "^1.0.5",
-    "gulp-imagemin": "^7.1.0",
+    "gulp-imagemin": "^8.0.0",
     "gulp-plumber": "^1.2.1",
     "gulp-pug": "^5.0.0",
     "gulp-pug-linter": "^1.5.0",
     "gulp-rename": "^2.0.0",
-    "gulp-sass": "^5.0.0",
+    "gulp-sass": "^5.1.0",
     "gulp-shorthand": "^1.1.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-stylelint": "^13.0.0",
-    "gulp-terser": "^2.0.1",
-    "gulp-w3c-html-validator": "^2.0.1",
+    "gulp-terser": "^2.1.0",
+    "gulp-w3c-html-validator": "^5.1.3",
     "gulp-webp": "^4.0.1",
-    "lighthouse": "^8.1.0",
-    "node-sass": "^6.0.1",
+    "lighthouse": "^9.6.3",
+    "node-sass": "^7.0.1",
     "pug": "^3.0.2",
-    "sass": "^1.36.0",
-    "stylelint": "^13.13.1",
-    "stylelint-config-htmlacademy": "^0.1.10",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-config-standard-scss": "^1.1.0",
-    "stylelint-scss": "^3.20.1"
+    "sass": "^1.54.0",
+    "stylelint": "^14.9.1",
+    "stylelint-config-htmlacademy": "^0.1.12",
+    "stylelint-config-standard": "^26.0.0",
+    "stylelint-config-standard-scss": "^5.0.0",
+    "stylelint-scss": "^4.3.0"
   },
   "dependencies": {
+    "caniuse-lite": "^1.0.30001369",
     "compression": "^1.7.4",
     "http2": "^3.3.7",
     "normalize.css": "^8.0.1"
   },
-  "types": "./types.d.ts"
+  "types": "./types.d.ts",
+  "type": "module"
 }
