@@ -1,16 +1,17 @@
-const gulp = require('gulp')
-const plumber = require('gulp-plumber')
-const pug = require('gulp-pug')
-const pugLinter = require('gulp-pug-linter')
-const htmlValidator = require ('gulp-w3c-html-validator')
-const bemValidator = require('gulp-html-bem-validator')
+import gulp from 'gulp'
+import plumber from 'gulp-plumber'
+import pug from 'gulp-pug'
+import pugLinter from 'gulp-pug-linter'
+import { htmlValidator } from 'gulp-w3c-html-validator'
+import bemValidator from 'gulp-html-bem-validator'
 
-module.exports = function pug2html(cb) {
+export function pug2html(cb) {
     return gulp.src('src/pages/*.pug')
         .pipe(plumber())
         .pipe(pugLinter({ reporter: 'default' }))
         .pipe(pug())
-        .pipe(htmlValidator())
+        .pipe(htmlValidator.analyzer())
+        .pipe(htmlValidator.reporter())
         .pipe(bemValidator())
         .pipe(gulp.dest('build'))
 }
